ADD: generating statically filesystem with 'cartridge create' files
