+++ cli/create/codegen/static/content/config/.gitignore
@@ -1,0 +1,22 @@
+.rocks
+.swo
+.swp
+CMakeCache.txt
+CMakeFiles
+cmake_install.cmake
+*.dylib
+*.idea
+__pycache__
+*pyc
+.cache
+.pytest_cache
+.vagrant
+.DS_Store
+*.xlog
+*.snap
+*.rpm
+*.deb
+*.tar.gz
+node_modules
+/tmp/*
+!/tmp/.keep
