+++ cli/create/codegen/static/content/config/.luacheckrc
@@ -1,0 +1,4 @@
+include_files = {'**/*.lua', '*.luacheckrc', '*.rockspec'}
+exclude_files = {'.rocks/', 'tmp/'}
+max_line_length = 120
+redefined = false
