+++ cli/create/codegen/static/content/dev/.cartridge.yml
@@ -1,0 +1,7 @@
+---
+# here you can specify default parametes for local running, such as
+
+# cfg: path-to-cfg-file
+# log-dir: path-to-log-dir
+# run-dir: path-to-run-dir
+# data-dir: path-to-data-dir
