+++ cli/create/codegen/static/content/test/api_test.lua
@@ -1,0 +1,12 @@
+local t = require('luatest')
+local g = t.group('integration_api')
+
+local helper = require('test.helper.integration')
+local cluster = helper.cluster
+
+g.test_sample = function()
+    local server = cluster.main_server
+    local response = server:http_request('post', '/admin/api', {json = {query = '{}'}})
+    t.assert_equals(response.json, {data = {}})
+    t.assert_equals(server.net_box:eval('return box.cfg.memtx_dir'), server.workdir)
+end
