+++ cli/create/codegen/static/content/test/helper.lua
@@ -1,0 +1,18 @@
+-- This file is required automatically by luatest.
+-- Add common configuration here.
+
+local fio = require('fio')
+local t = require('luatest')
+
+local helper = {}
+
+helper.root = fio.dirname(fio.abspath(package.search('init')))
+helper.datadir = fio.pathjoin(helper.root, 'tmp', 'db_test')
+helper.server_command = fio.pathjoin(helper.root, 'init.lua')
+
+t.before_suite(function()
+    fio.rmtree(helper.datadir)
+    fio.mktree(helper.datadir)
+end)
+
+return helper
