+++ cli/create/codegen/static/content/test/unit.lua
@@ -1,0 +1,9 @@
+local t = require('luatest')
+
+local shared = require('test.helper')
+
+local helper = {shared = shared}
+
+t.before_suite(function() box.cfg({work_dir = shared.datadir}) end)
+
+return helper
