--- cli/create/codegen/static/doc.go
@@ -1,1 +1,0 @@
-package static
